Move common components to packages directory
