+++ src/blocks/column/edit.js
@@ -15,14 +15,11 @@ import icons from '../../icons';
 import MeasurementControls from '../../measurement-control';
 import classnames from 'classnames';
 import debounce from 'lodash/debounce';
-import { PopColorControl, ResponsiveAlignControls, BoxShadowControl } from '@kadence/components';
-import KadenceBackgroundControl from '../../components/background/background-control';
+import { PopColorControl, VerticalAlignmentIcon, KadenceRadioButtons, ResponsiveAlignControls, BoxShadowControl, BackgroundControl as KadenceBackgroundControl } from '@kadence/components';
 import { KadenceColorOutput } from '@kadence/helpers';
 import KadenceRange from '../../components/range/range-control';
 import ResponsiveMeasuremenuControls from '../../components/measurement/responsive-measurement-control';
-import KadenceRadioButtons from '../../components/common/kadence-radio-buttons';
 import SmallResponsiveControl from '../../components/responsive/small-responsive-control';
-import VerticalAlignmentIcon from '../../components/common/vertical-align-icons';
 import ResponsiveRangeControls from '../../components/range/responsive-range-control';
 import KadencePanelBody from '../../components/KadencePanelBody';
 import URLInputControl from '../../components/links/link-control';
