+++ src/blocks/icon/edit.js
@@ -14,10 +14,9 @@ import times from 'lodash/times';
 import map from 'lodash/map';
 import IconControl from '../../components/icons/icon-control';
 import IconRender from '../../components/icons/icon-render';
-import { PopColorControl } from '@kadence/components';
+import { PopColorControl, VerticalAlignmentIcon } from '@kadence/components';
 import { KadenceColorOutput } from '@kadence/helpers';
 import StepControl from '../../step-control';
-import VerticalAlignmentIcon from '../../components/common/vertical-align-icons';
 import KadencePanelBody from '../../components/KadencePanelBody';
 import URLInputControl from '../../components/links/link-control';
