+++ src/blocks/testimonials/edit.js
@@ -26,8 +26,7 @@ import TypographyControls from '../../components/typography/typography-control';
 import MeasurementControls from '../../measurement-control';
 import IconControl from '../../components/icons/icon-control';
 import IconRender from '../../components/icons/icon-render';
-import KadenceMediaPlaceholder from '../../kadence-media-placeholder';
-import { PopColorControl, WebfontLoader } from '@kadence/components';
+import { PopColorControl, WebfontLoader, KadenceMediaPlaceholder } from '@kadence/components';
 import { KadenceColorOutput } from '@kadence/helpers';
 import ResponsiveRangeControls from '../../components/range/responsive-range-control';
 import ResponsiveMeasuremenuControls from '../../components/measurement/responsive-measurement-control';
