+++ src/packages/components/src/background-control/editor.scss
