+++ src/packages/components/src/drop-shadow-control/index.js
@@ -6,17 +6,13 @@
 /**
  * Import Externals
  */
- import { PopColorControl } from '@kadence/components';
+import { PopColorControl } from '@kadence/components';
 /**
  * Internal block libraries
  */
 import { __ } from '@wordpress/i18n';
-const {
-	Component,
-} = wp.element;
-const {
-	ToggleControl,
-} = wp.components;
+import { Component } from '@wordpress/element';
+import { ToggleControl } from '@wordpress/components';
 /**
  * Build the DropShadow controls
  * @returns {object} DropShadow settings.
