+++ src/packages/components/src/index.js
@@ -11,3 +11,12 @@ export { default as HeadingLevelIcon } from './heading-level-icon';
 export { default as BoxShadowControl } from './box-shadow-control';
 export { default as DynamicImageControl } from './dynamic-image-control';
 export { default as DynamicTextControl } from './dynamic-text-control';
+export { default as DynamicBackgroundControl } from './dynamic-background-control';
+export { default as DropShadowControl } from './drop-shadow-control';
+export { default as BackgroundControl } from './background-control';
+export { default as KadenceFocalPicker } from './focal-picker';
+export { default as KadenceRadioButtons } from './common/radio-buttons';
+export { default as KadenceMediaPlaceholder } from './common/media-placeholder';
+export { default as ImageSizeControl } from './common/image-size-controls';
+export { default as VerticalAlignmentIcon } from './common/vertical-align-icon';
+export { default as KadenceImageControl } from './common/image-controls';
