+++ src/packages/helpers/src/index.js
@@ -4,3 +4,4 @@ export { default as KadenceColorOutput } from './kadence-color-output';
 export { default as DeprecatedKadenceColorOutput } from './deprecated-kadence-color-output';
 export { default as fetchJson } from './fetch-json';
 export { default as capitalizeFirstLetter } from './capital-first';
+export { default as KadenceTryParseJSON } from './parse-json';
