+++ src/plugins/prebuilt-library/cloud-connect.js
@@ -41,7 +41,7 @@ import LazyLoad from 'react-lazy-load';
 /**
  * Internal dependencies
  */
- import KadenceTryParseJSON from '../../components/common/parse-json'
+ import { KadenceTryParseJSON } from '@kadence/helpers'
 /**
  * Internal block libraries
  */
@@ -183,7 +183,7 @@ class CloudConnect extends Component {
 						</Button>
 					</div>
 					<div className="kb-connection-info">
-						{ __( 'Learn about connecting libraries at:', 'kadence-blocks' ) + ' ' } 
+						{ __( 'Learn about connecting libraries at:', 'kadence-blocks' ) + ' ' }
 						<ExternalLink href={ 'https://www.kadencewp.com/kadence-cloud/?utm_source=in-app&utm_medium=kadence-blocks&utm_campaign=design-library' }>{ __( 'Kadence Cloud', 'kadence-blocks' ) }</ExternalLink>
 					</div>
 					</Fragment>
