+++ src/plugins/prebuilt-library/cloud-library.js
@@ -19,7 +19,7 @@ import {
 	withDispatch,
 } from '@wordpress/data';
 import { rawHandler } from '@wordpress/blocks';
-import { 
+import {
 	Component,
 	Fragment,
 } from '@wordpress/element';
@@ -46,7 +46,7 @@ import { __, sprintf } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import KadenceTryParseJSON from '../../components/common/parse-json'
+import { KadenceTryParseJSON } from '@kadence/helpers'
 
 
 class CloudSections extends Component {
@@ -232,7 +232,7 @@ class CloudSections extends Component {
 		} );
 		const sideCatOptions = Object.keys( categoryItems ).map( function( key, index ) {
 			return { value: ( 'category' === key ? 'all' : key ), label: ( 'category' === key ? __( 'All', 'kadence-blocks' ) : categoryItems[key] ) }
-		} );	
+		} );
 		const control = this;
 		let breakpointColumnsObj = {
 			default: 5,
@@ -383,7 +383,7 @@ class CloudSections extends Component {
 									{ __( 'Error, Unable to access library database, please try re-syncing', 'kadence-blocks' ) }
 								</h2>
 								<div style={ { textAlign:'center' } }>
-									<Button 
+									<Button
 										className="kt-reload-templates"
 										icon={ update }
 										onClick={ () => this.reloadTemplateData() }
