+++ src/plugins/prebuilt-library/section-library.js
@@ -26,7 +26,7 @@ import {
 	withDispatch,
 } from '@wordpress/data';
 import { rawHandler } from '@wordpress/blocks';
-import { 
+import {
 	Component,
 	Fragment,
 } from '@wordpress/element';
@@ -53,7 +53,7 @@ import { __, sprintf } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
- import KadenceTryParseJSON from '../../components/common/parse-json'
+import { KadenceTryParseJSON } from '@kadence/helpers'
 
 /**
  * Prebuilt Sections.
@@ -374,7 +374,7 @@ class PrebuiltSections extends Component {
 									{ __( 'Error, Unable to access library database, please try re-syncing', 'kadence-blocks' ) }
 								</h2>
 								<div style={ { textAlign:'center' } }>
-									<Button 
+									<Button
 										className="kt-reload-templates"
 										icon={ update }
 										onClick={ () => this.reloadTemplateData() }
